+++ GPTifier/include/cli.hpp
@@ -3,6 +3,5 @@
 namespace cli {
 
 void exit_on_failure();
-bool get_opts_files_list(int argc, char **argv);
 
 } // namespace cli
